requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
kafka-python==2.0.2
structlog==23.1.0
konlpy==0.6.0
nltk==3.8.1 